--- src/components/LogorSign/LogorSign.css
@@ -1,77 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Roboto:300);
-
-.flex{ 
-    display: flex;
-    padding: 5px;
-    
-}
-.formContainer{ 
-    position: relative;
-    z-index: 1;
-    background: #FFFFFF;
-    background-image: none;
-    max-width: 50%;
-    margin: 100px auto 100px auto;
-    padding: 45px;
-    text-align: center;
-    box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.2), 0 5px 5px 0 rgba(0, 0, 0, 0.24);
-    border-radius: 10px;
-
-}
-
-.smallContainer { 
-
-    background-color: #FFFFFF;
-    background-image: none;
-    padding: 2px;
-    
-}
-.test { 
-    border: #08ac4b solid 2px;
-    padding: 10px;
-
-}
-.logorsign-input { 
-    margin: 5px;
-    min-width: 50px;
-    font-family: "Roboto";
-    border-radius: 5px;
-    height: auto;
-    background-color: #f2f2f2;
-    background-image: none;
-    border: none;
-    padding: 10px;
-}
-
-.switchButton { 
-    font-family: "Roboto", sans-serif;
-    text-transform: uppercase;
-    background-color: #328f8a;
-    background-image: linear-gradient(45deg,#328f8a,#08ac4b);
-    width: 50%;
-    border: none;
-    border-radius: 5px;
-    margin: 5px 0 0 0;
-    padding: 10px;
-    color: #FFFFFF;
-    font-size: 14px;
-    -webkit-transition: all 0.3 ease;
-    transition: all 0.3 ease;
-    cursor: pointer;
-}
-.loginButton { 
-    border-radius: 5px;
-    font-family: "Roboto", sans-serif;
-    text-transform: uppercase;
-    background-color: #328f8a;
-    background-image: linear-gradient(45deg,#328f8a,#08ac4b);
-    width: 50%;
-    border: none;
-    margin: 5px 0 0 0;
-    padding: 10px;
-    color: #FFFFFF;
-    font-size: 14px;
-    -webkit-transition: all 0.3 ease;
-    transition: all 0.3 ease;
-    cursor: pointer;
-}
